In / Out Page Source Code Added
1. Stock in /out function
2. check stock function

+++ FactoryManagementSoftware/FactoryManagementSoftware/BLL/itemBLL.cs
@@ -11,8 +11,8 @@ namespace FactoryManagementSoftware.BLL
         public string item_code { get; set; }
         public string item_name { get; set; }
         public string item_cat { get; set; }
-        public int item_qty { get; set; }
-        public int item_ord { get; set; }
+        public float item_qty { get; set; }
+        public float item_ord { get; set; }
         public DateTime item_added_date { get; set; }
         public int item_added_by { get; set; }
         public DateTime item_updtd_date { get; set; }
